move slippage to trade widget
