+++ src/elements/settings/SlippageSettings.tsx
@@ -1,0 +1,77 @@
+import { useDispatch } from 'react-redux';
+import { setSlippageTolerance } from 'store/user/user';
+import { useAppSelector } from 'store';
+import { useState } from 'react';
+
+export const SlippageSettings = () => {
+  const currentSlippage = useAppSelector<number>(
+    (state) => state.user.slippageTolerance
+  );
+  const slippages = [0.001, 0.005, 0.01];
+  const [customSlippage, setCustomSlippage] = useState(
+    slippages.includes(currentSlippage)
+      ? ''
+      : (currentSlippage * 100).toString()
+  );
+
+  const dispatch = useDispatch();
+
+  const normalizedSlippage = Number(customSlippage) / 100;
+
+  return (
+    <div className="space-y-15 text-black-low dark:text-white-low">
+      <div>Slippage Tolerance</div>
+      <p className={'text-12'}>
+        Your transaction will rever if the price changes unfavorably by more
+        then this percentage.
+      </p>
+      <div className="flex justify-between space-x-6">
+        {slippages.map((slippage) => (
+          <button
+            key={slippage}
+            onClick={() => {
+              dispatch(setSlippageTolerance(slippage));
+              setCustomSlippage('');
+            }}
+            className={`w-full border border-silver dark:border-grey text-black dark:text-white rounded-[12px] text-12 p-8 ${
+              currentSlippage === slippage ? 'bg-fog dark:bg-grey' : ''
+            }`}
+          >
+            +{slippage * 100}%
+          </button>
+        ))}
+        <span
+          className={`flex items-center border border-silver dark:border-grey rounded-[12px] pr-5 text-12 text-black dark:text-white ${
+            currentSlippage === normalizedSlippage &&
+            !slippages.includes(currentSlippage)
+              ? 'bg-fog dark:bg-grey'
+              : 'bg-white dark:bg-black'
+          }`}
+        >
+          <input
+            type="text"
+            className={`w-[60px] border-none outline-none text-center ${
+              currentSlippage === normalizedSlippage &&
+              !slippages.includes(currentSlippage)
+                ? 'bg-fog dark:bg-grey'
+                : 'bg-white dark:bg-black'
+            }`}
+            value={customSlippage}
+            onChange={(event) => setCustomSlippage(event.target.value)}
+            onBlur={() => {
+              if (
+                customSlippage.trim() !== '' &&
+                !isNaN(Number(customSlippage))
+              )
+                dispatch(setSlippageTolerance(normalizedSlippage));
+
+              slippages.includes(currentSlippage) && setCustomSlippage('');
+            }}
+            placeholder="Custom"
+          />
+          %
+        </span>
+      </div>
+    </div>
+  );
+};
