+++ src/elements/swapHeader/SwapHeader.tsx
@@ -1,6 +1,9 @@
 import { classNameGenerator } from 'utils/pureFunctions';
 import 'elements/swapHeader/SwapHeader.css';
 import { SwapSwitch } from 'elements/swapSwitch/SwapSwitch';
+import { PopoverV3 } from 'components/popover/PopoverV3';
+import { ReactComponent as IconSettings } from 'assets/icons/settings.svg';
+import { SlippageSettings } from 'elements/settings/SlippageSettings';
 
 interface SwapHeaderProps {
   isLimit: boolean;
@@ -32,7 +35,16 @@ export const SwapHeader = ({ isLimit, setIsLimit }: SwapHeaderProps) => {
           </button>
         </div>
 
-        <SwapSwitch />
+        <div className={'flex items-center space-x-20'}>
+          <PopoverV3
+            buttonElement={() => (
+              <IconSettings className={'w-24 cursor-pointer'} />
+            )}
+          >
+            <SlippageSettings />
+          </PopoverV3>
+          <SwapSwitch />
+        </div>
       </div>
     </div>
   );
